Work around circular dependency involving babel-core.
Strange as it may sound, calling require("babel-core") for the first time
can potentially result in calling the retrieveSourceMap function if the
source-map-support package is active, and any stack traces need to be
rewritten (which happens whenever evaluating error.stack).

As a symptom of this circular dependency, require("babel-core") could be
called again before returning the first time, thus exposing an incomplete
object with no transformFromAst method.

With this commit, if require("babel-core").transformFromAst is not
defined, retrieveSourceMap just returns null to indicate it couldn't
retrieve a source map, which is a perfectly allowable behavior that
shouldn't cause any problems.