Enable dead code elimination in minifier-js to fix #9260.
The react-dom@16 npm package recently got stricter about how it checks
that dead code elimination is enabled. Specifically, it no longer just
checks that process.env.NODE_ENV is replaced with "production" but also
requires conditional branches to be completely removed if the replacement
makes them unreachable:
https://github.com/facebook/react/blob/master/packages/react-dom/npm/index.js

The minifier-js package recently had its minor version bumped for Meteor
1.6, so I think it's safe to enable dead code elimination in a patch
version update. If we bumped the minor version again (to 2.3.0), this
version of minifier-js would not be usable with Meteor 1.6, and would
probably have to wait for Meteor 1.6.1.

Bumping the patch version of standard-minifier-js is also necessary
because it uses minifier-js as a compiler plugin, and would otherwise
continue to bundle the old minifier-js plugin.