Move 'cordova-lib' back to CORDOVA_DEV_BUNDLE_VERSIONS for auto install.
While 'cordova-lib' used to live in 'dev-bundle-tool-package.js' where
it was pre-bundled into the "dev bundle", that is no longer the case.

It is now automatically installed, on demand, when Cordova is used.

This follows up on meteor/meteor#9213, which added it back to the
'dev-bundle-tool-package.js' and updates it in the new location.

Ref: https://github.com/meteor/meteor/pull/9213
Ref: https://github.com/meteor/meteor/commit/073f2410a685f1de87f70dff6a097b1eea2712b1

cc @skirunman.