Removed cordova-plugin-console dependency
As stated on [https://github.com/apache/cordova-plugin-console](cordova-plugin-console):

This plugin is no longer being worked on as the functionality provided by this plugin is now included in cordova-ios 4.5.0 or greater, and support is already built in to cordova-windows > 5.0.0. You should remove this plugin from your applications.

Please file issues for this plugin against their respective platforms (cordova-ios, cordova-windows).