Update variable on outdated documentation.
This "long form" documentation is not really actively updated anymore, however, it's harmless to change this variable to something that actually works now.  In this case, the variable has to be something which is parsable by `JSON.parse`.

Addresses the confusion caused by this doc, as outlined in https://github.com/meteor/meteor/issues/9395#issuecomment-345848605.