Update README.md to incorporate BrowserStack badge.
Hopefully, without too much effort, it will be easy to reintegrate much of
the automated BrowserStack testing we (mostly) already had in place!  In the
near future, this could be helpful for ensuring we're not over/under-shipping
polyfills to browsers.

Hopefully we can keep this green, though it's not clear to me at the
moment what additional changes I'll need to make to ensure that.  For
now, badge!