Changed `failedTests` to live on the `testList`.
The test itself already keeps track of its failure, and the test list
tracks other concepts of grouped tests, such as "filtered tests",
"all tests", etc. (in a manner consistent with what would be expected
from a "test list", so I believe it makes sense for it to also maintain a
collection of "failed tests"... in list form, of course!

This also prevents needing to pass back a failed test and accumulating
it in other means, and takes advantage of the existing `notifyFailed`
facility.