Test which ensures that the global `_` isn't compromised.
As demonstrated in https://github.com/meteor/meteor/issues/9276.

This test wouldn't have caught the regression in the previous solution
since the lack of a TTY in the `self-test` test harness caused the tests
themselves to take the path through `shell-server`'s `evaluateAndExit`
logic, which didn't use the `global` scope in the same way as the
interactive shell.  That is no longer the case as of e0682c553ddf0a9.