Unescape ampersands in oauth redirect URLs (#9484)
When using the `oauth` package "redirect" `loginStyle`, the
redirect URL is escaped before it is embedded in the HTML
page that is used to finalize the oauth process. This
escaped redirect URL is then assigned directly to
`window.location` to finalize the redirect. Browsers can
properly handle most of the escaped URL components, with
the exception of HTML entity based ampersands: `&amp;`.
The `&amp;`'s are left in the redirect URL after the
redirect has completed, leading to broken redirects in
some cases.

This commit makes sure `&amp;`'s are converted back to
`&`'s before the redirect URL is assigned to
`window.location`.

Fixes #9279