Test improvements to async_hooks support in node-fibers.
The async_hooks API seems not entirely thread-aware, which can cause
problems when fibers are used, since fibers behave a lot like threads from
the perspective of Node/V8.

The author of node-fibers (@laverdet) has found a way to save/restore the
async_hooks stack, and wants feedback on the effectiveness of the changes:
https://github.com/laverdet/node-fibers/issues/357#issuecomment-356175889

The relevant commit (not yet published to npm):
https://github.com/laverdet/node-fibers/commit/c6fec5c378dc94f290382a44d60802541c0ac9c8

If these changes work, I hope to turn them into a proper Node PR, but in
the meantime it seems useful to run Meteor's test suite against them.