Go back to writing .meteor-portable-2.json files asynchronously.
This reverts commit 4e4e204ab0891ac5a8f90820955147e35a302d9c.

This commit caused a strange regression in reliability of the Windows
dynamic-import self-test, which may be an indication of a deeper problem,
so it seems safest to revert this change for now.

In case empty .meteor-portable-2.json files are written, I've added an
additional check that the cached JSON value is a boolean.