Revert "Update `code` submodule to latest `master` for 1.6.1 docs."
This reverts commit e895e11349f187c4ba004382077eca4f582f7406.

Since apparently there are some issues with the "apibox" calls in
regards to some of our recently modernized code and their associated
jsdoc declarations.  Specifically, the following API box calls are being
problematic when running `npm start` in this repository (and thus when
deploying too):

* Meteor.status
* check

The failures are happening when jsdoc is ran on the repository, which builds
the data/*.json files which are consumed by the (custom) apibox method.

I'll need to look into this further but, for now, I'll revert this commit
and bump the `code` submodule to a branch on `meteor/meteor` with just
the updated History.md in order to get the changelog updated on docs
before I log for the weekend.

cc @benjamn @hwillson