Special branch for to fix the docs for 1.6.1.
This commit represents the hash which the docs were previously on
(release-1.6.0.1) but with the exact History.md for the latest 1.6.x in
order to properly generate the "Changelog" which is shown on the
sidebar.

This is necessary because some of the jsdoc declarations for changes in
the 1.6.1 release affected the docs generation, which are expecting
particular declarations which are apparently missing.

These are fixed on 1.8, but persists for 1.7 and 1.6.