Export `check` as a named function declaration for JSDoc.
The change made in a30f42c4acabb7591cdde61a1eaeac811610ba3e switched
from CommonJS to ECMAScript export notation.

It seems JSDoc isn't smart enough to make the same association as it did with
the previous notation as it does for an anon. function expression `export`-ed
as a `const`ant.  We could annotate this with `@function check`, but it seems
reasonable to just export a function declaration directly, which JSDoc will
understand.

This, along with ababb18aef6d75055f1f460ae4709c2ef26c4836, corrects the
failure of docs generation originally discovered in:
https://github.com/meteor/docs/commit/8f7ceb5de5bf9bdd5935e3447164ce8f24cece7b.

Refs: https://github.com/meteor/meteor/pull/9593