Add misssing JSDoc declarations for Mongo.Collection#upsert.
In the conversion to ECMAScript classes in
https://github.com/meteor/meteor/commit/6fcf190bb489db19, it appears that
JSDoc `@method`, `@instance` and `@memberof` declarations were already
in place for the (outgoing) `Mongo.Collection.prototype.{foo}` expressions,
with the exception of `upsert`.  The omission caused docs generation failure
with the new ES syntax since, presumably, it was less possible to be inferred
from the `MCp*` notation.

Additionally, this changes the casing of `@memberOf` to `@memberof`,
mainly because:

* The JSDocs use this notation.
* The data JSON outputted by JSDoc lowercases property keys, including
  cases where `memberOf` was capitalized, making it less clear that
  there is a 1:1 mapping.
* My editor refuses to syntax-highlight jsdoc declarations with [A-Z],
  possilby because of the above reasons.