Run iOS tests automatically on TravisCI.
By using TravisCI's Mac testing environment, which comes fully-equipped
with Xcode and iOS simulators, we can run the test suite which @hwillson
revived in #46, automatically.

While many of the tests are failing, there are some passing, which seems
to prove that the testing framework is functioning properly, though
doesn't provide much assurance that we'll be able to fix the tests (though
it's likely, and I'm hopeful, that we can!).

Despite hesitance to do so, I've specifically pinned the iOS simulator to
use iOS 11.2 since the `cordova-paramedic` npm tries to obtain the logs from
the simulator by looking through the output of `instruments -s devices`, though
the `cordova` npm is only aware of minor versions in iOS releases, not patch
versions (e.g 11.0.1, which it classifies as 11.0).  Since Travis is
automatically updating to patch versions of iOS, presumably in an effort
to solve the security problems or deficiencies the initial releases had,
this seems necessary for the time being.

Anyone should bump the simulator OS version as necessary/beneficial.  We could
potentially fix this in our own fix to `cordova-paramedic`, which I've
already had to fork for unrelated reasons, but I haven't taken the time
to investigate whether this is reasonable.

This also adds `ios-deploy` and `ios-sim` as `devDependencies` for testing, and
removes them from their previous installation location in `.travis.yml`,
which makes it possible to run `npm test` locally.