Add a script to generate a Netlify _redirects file.
The redirects in this _redirects file are generated automatically from
the `versions` configuration array in the `_config.yml` and will produce
"proxy" redirects to go to the correct Netlify deployment for the
appropriate version (which are based on branch names).

I'll be honest, this may fail, but putting this on `master` is the
easiest way to test!