A special commit on a special branch to fix Meteor 1.3.5.x docs.
This branch (docs/v1.3.5) can be deleted if the v1.3.5 docs are ever
sunset, but should be kept otherwise.

This commit backports a special fix onto a special 1.3.5.x docs branch
which will be referenced in the `code` submodule on the
meteor/docs@version-1.3.5 branch of the docs repo.

That commit was originally intended to fix exactly this problem, but it
never got backported prior to this, and apparently the deploys have been
failing on 1.3.5 for quite some time now.  Voi!